Fix to uint256

+++ contracts/libraries/VaultLifecycleYearn.sol
@@ -32,7 +32,7 @@ library VaultLifecycleYearn {
         address USDC;
         address currentOption;
         uint256 delay;
-        uint16 lastStrikeOverride;
+        uint256 lastStrikeOverride;
         uint256 overriddenStrikePrice;
     }
